Add the Sensor class and the corresponding test class
A basic implementation of the sensor basic class has been added. In the future it would probably be advised to make this class abstract with some functions that will be implemented by the simulations (i.e. what fields should be saved by the system).
The corresponding test class has been added with basic tests (that guided me into creating this class).