Make the Sensor class abstract and add a method #44
The sensor class has been made abstract because an abstract method has been added. A documentation of the method can be found in the code, and it is responsible for telling the system what columns should be allocated in the database for this sensor.
The tests have been changed as well, and a new one has been created.