Sensors: Start to handle deadlock bug
Added a type to Sensor objects.
Opened a new thread from SensorsHandler in case the connecting sensor is
an interactive sensor. This thread should wait for the sensor to connect
via it instructions connection.
Had to use many lambda expressions in order to pipe information backwards
from the threads.

Issue: #143