Testing: Add test and find a bug
Added test to SensorLocation enum, using the parameterized testing method.
Adding this test changed the order in which all tests being run and so I
discovered that when the GUI test is running before the tests in
SensorsManagerTest.java, it causes them to fail.
The GUI test is now ignored with a proper documentation.

Issue: #138