Rewrote the SensorApi to work with our new changes.
For the developer, the API will look very similar to the previous one.
Under the hood, it is completely redesigned. It uses the FileSystem in order to listen for new messages, and to send instructions to sensors.
A big change is that there are no more exceptions of type SensorNotFound. The applications can now be loaded even if the sensors are not connected. This means that the API maintains command buffers from the user. And when a sensor is found, the API will push the buffered actions to the file system.
The rest of the changes will be explained in the future and hopefully in the Wiki pages