Operation modes: Themes and CSS
The developer theme is now defined by a CSS file (The content of the
description in the home and the removal of the user information and
sensors tabs are still being controlled from the code).
The plus button and the drag&drop area are now hidden in Developer Mode.
The dark theme is almots complete.

Issue: #189